# backend/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Counter-Inscriptions Minter API
  version: 2.0.0
  description: |
    Mint any MIME type as an Ordinals inscription via the Counterparty protocol (taproot encoding).

    ## Two transactions per inscription
    A Counterparty taproot inscription is a **commit** transaction plus a **reveal** transaction:

    1. `tx.psbt` — the unsigned commit as a base64 PSBT. The connected wallet signs and broadcasts it.
    2. `tx.signed_reveal_rawtransaction` — the reveal, already signed by the Counterparty node with an
       ephemeral key. Broadcast it with `POST /api/broadcast` **after** the commit. The inscription content
       is not on-chain until the reveal confirms.

    ## Who receives what
    Counterparty issuance semantics: the **signer** (`walletAddress`) pays fees, receives the issued units and
    the inscribed sat. `destinationWallet` only transfers asset *ownership* (`transfer_destination`). To move
    units to other wallets use `POST /api/send-batch` (one MPMA transaction) once the issuance has confirmed.

    ## Chunking
    Files larger than 350KB are split into consecutive numeric assets (`A<n>`, `A<n+1>`, …), one commit/reveal
    pair each. Named assets cannot be chunked. Reassembly is done by the viewer.
servers:
  - url: http://localhost:3001/api
    description: Local development server
  - url: /api
    description: Same origin (nginx proxy in Docker)
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
        details: { type: string }
        hint: { type: string }
paths:
  /mint:
    post:
      tags: [Minting]
      summary: Compose inscription issuance(s) for the connected wallet (auto-chunks large files)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, asset, walletAddress]
              properties:
                file:
                  type: string
                  format: binary
                  description: File to inscribe (max 50MB). Chunked above 350KB.
                asset:
                  type: string
                  example: "A1000000000000000000"
                  description: Numeric `A` + 17–20 digits (free) or 4–12 uppercase letters not starting with A (0.5 XCP).
                walletAddress:
                  type: string
                  description: Connected wallet. Signs the commit, pays fees, receives the units and the inscribed sat.
                destinationWallet:
                  type: string
                  description: Optional. Becomes the asset owner via `transfer_destination`. Does NOT receive the units.
                mime_type:
                  type: string
                  description: Override auto-detected MIME type.
                quantity:
                  type: integer
                  default: 1
                sat_per_vbyte:
                  type: number
                  default: 2.01
                encoding:
                  type: string
                  enum: [taproot]
                  default: taproot
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [asset, walletAddress, description]
              properties:
                asset: { type: string }
                walletAddress: { type: string }
                description: { type: string, description: Hex-encoded content (legacy clients) }
                mime_type: { type: string }
                sat_per_vbyte: { type: number }
      responses:
        '200':
          description: Composed transactions, one per chunk
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  asset: { type: string }
                  chunk_assets:
                    type: array
                    items: { type: string }
                  signer: { type: string }
                  holder: { type: string, description: Address that will hold the issued units (= signer) }
                  owner: { type: string, description: Asset owner after issuance }
                  total_chunks: { type: integer }
                  file_size_bytes: { type: integer }
                  mime_type: { type: string }
                  transactions:
                    type: array
                    items:
                      type: object
                      properties:
                        asset: { type: string }
                        chunk: { type: integer }
                        total_chunks: { type: integer }
                        chunk_bytes: { type: integer }
                        tx: { $ref: '#/components/schemas/ComposedTx' }
                  next_steps:
                    type: array
                    items: { type: string }
        '400': { description: Missing/invalid fields or a named asset that would need chunking, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '402': { description: Insufficient funds, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: Asset name already exists, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '422': { description: Unsupported MIME type, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '503': { description: Counterparty server unreachable }
  /mint-batch:
    post:
      tags: [Minting]
      summary: Airdrop step 1 — one issuance of quantity = number of wallets
      description: |
        Composes a single inscription issuance from `walletAddress` with `quantity` equal to the number of
        distinct `destinationWallets`. After it confirms, call `/send-batch` to deliver one unit to each wallet.
        File must be ≤350KB.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, asset, walletAddress, destinationWallets]
              properties:
                file: { type: string, format: binary }
                asset: { type: string }
                walletAddress: { type: string }
                destinationWallets: { type: string, description: Comma-separated addresses, example: "bc1p…,bc1p…" }
                mime_type: { type: string }
                sat_per_vbyte: { type: number, default: 2.01 }
      responses:
        '200':
          description: Composed issuance
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  asset: { type: string }
                  signer: { type: string }
                  quantity: { type: integer }
                  destinations: { type: array, items: { type: string } }
                  mime_type: { type: string }
                  file_size_bytes: { type: integer }
                  tx: { $ref: '#/components/schemas/ComposedTx' }
                  next_steps: { type: array, items: { type: string } }
        '400': { description: Missing fields }
        '409': { description: Asset name already exists }
        '413': { description: File too large for a single chunk }
  /send-batch:
    post:
      tags: [Minting]
      summary: Airdrop step 2 — one MPMA send to many wallets
      description: The asset must be confirmed and held by `walletAddress`; otherwise Counterparty reports insufficient funds.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [walletAddress, asset, destinationWallets]
              properties:
                walletAddress: { type: string }
                asset: { type: string }
                destinationWallets:
                  oneOf:
                    - { type: array, items: { type: string } }
                    - { type: string, description: Comma-separated }
                quantity_each: { type: integer, default: 1 }
                sat_per_vbyte: { type: number, default: 2.01 }
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                walletAddress: { type: string }
                asset: { type: string }
                destinationWallets: { type: string }
      responses:
        '200':
          description: Composed MPMA send (no reveal — sign `tx.psbt` and broadcast)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  asset: { type: string }
                  signer: { type: string }
                  destinations: { type: array, items: { type: string } }
                  quantity_each: { type: integer }
                  tx: { $ref: '#/components/schemas/ComposedTx' }
        '402': { description: Insufficient funds (issuance not yet confirmed?) }
  /broadcast:
    post:
      tags: [Broadcast]
      summary: Broadcast a signed transaction hex (use for the reveal)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [signedhex]
              properties:
                signedhex: { type: string, description: Signed transaction hex }
      responses:
        '200':
          description: Broadcast accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  txid: { type: string }
        '400': { description: Not a hex transaction }
        '500': { description: Node rejected the transaction (details in `error`) }
  /upload:
    post:
      tags: [Lookup]
      summary: Pre-flight analysis — MIME, size, chunk count, hex preview
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200':
          description: Analysis
          content:
            application/json:
              schema:
                type: object
                properties:
                  originalName: { type: string }
                  mimeType: { type: string }
                  sizeBytes: { type: integer }
                  totalChunks: { type: integer }
                  needsChunking: { type: boolean }
                  hexPreview: { type: string }
  /asset/{asset}:
    get:
      tags: [Lookup]
      summary: Check whether an asset name is valid and still available
      parameters:
        - { name: asset, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Availability
          content:
            application/json:
              schema:
                type: object
                properties:
                  asset: { type: string }
                  valid: { type: boolean }
                  available: { type: boolean }
                  existing: { type: object, nullable: true }
  /tx/{hash}:
    get:
      tags: [Lookup]
      summary: Counterparty view of a transaction (`result` is null until parsed)
      parameters:
        - { name: hash, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Counterparty transaction record or null }
  /balance/{address}:
    get:
      tags: [Lookup]
      summary: Counterparty balances for an address
      parameters:
        - { name: address, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: "Proxied /v2/addresses/{address}/balances" }
...